rubocop-fix long class
