+++ src/.rubocop_todo.yml
@@ -45,6 +45,7 @@ Lint/ShadowingOuterLocalVariable:
 Lint/SuppressedException:
   Exclude:
     - 'app/controllers/shifts_controller.rb'
+    - 'app/controllers/csv_shifts_controller.rb'
     - 'test/controllers/api/reservations_controller_test.rb'
 
 # Offense count: 12
@@ -53,6 +54,7 @@ Lint/SuppressedException:
 Lint/UnusedBlockArgument:
   Exclude:
     - 'app/controllers/shifts_controller.rb'
+    - 'app/controllers/csv_shifts_controller.rb'
     - 'app/models/concerns/shift_build_module.rb'
     - 'app/models/reservation.rb'
     - 'app/models/shift.rb'
@@ -86,6 +88,7 @@ Lint/UselessAssignment:
     - 'app/controllers/menus_controller.rb'
     - 'app/controllers/roles_controller.rb'
     - 'app/controllers/shifts_controller.rb'
+    - 'app/controllers/csv_shifts_controller.rb'
     - 'app/controllers/skills_controller.rb'
     - 'app/controllers/staffs_controller.rb'
     - 'app/controllers/stores_controller.rb'
@@ -108,7 +111,7 @@ Metrics/BlockLength:
 # Offense count: 3
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 112
+  Max: 120
 
 # Offense count: 1
 # Configuration parameters: IgnoredMethods.
@@ -288,6 +291,7 @@ Style/IfUnlessModifier:
   Exclude:
     - 'app/controllers/reservations_controller.rb'
     - 'app/controllers/shifts_controller.rb'
+    - 'app/controllers/csv_shifts_controller.rb'
     - 'app/controllers/staffs_controller.rb'
     - 'app/helpers/application_helper.rb'
     - 'app/helpers/customers_helper.rb'
@@ -389,6 +393,7 @@ Style/PreferredHashMethods:
     - 'app/controllers/observations_controller.rb'
     - 'app/controllers/reservations_controller.rb'
     - 'app/controllers/shifts_controller.rb'
+    - 'app/controllers/csv_shifts_controller.rb'
     - 'app/helpers/shifts_helper.rb'
     - 'app/models/shift.rb'
 
@@ -403,6 +408,7 @@ Style/RedundantBegin:
     - 'app/controllers/menus_controller.rb'
     - 'app/controllers/roles_controller.rb'
     - 'app/controllers/shifts_controller.rb'
+    - 'app/controllers/csv_shifts_controller.rb'
     - 'app/controllers/skills_controller.rb'
     - 'app/controllers/staffs_controller.rb'
     - 'app/controllers/stores_controller.rb'
@@ -521,7 +527,6 @@ Style/SymbolLiteral:
 Style/SymbolProc:
   Exclude:
     - 'app/controllers/customers/duplicate_controller.rb'
-    - 'app/controllers/shifts_controller.rb'
     - 'app/mailers/customer_mailer.rb'
     - 'app/mailers/reservation_mailer.rb'
     - 'app/models/concerns/shift_build_module.rb'
