+++ src/app/controllers/shifts_controller.rb
@@ -2,7 +2,6 @@ class ShiftsController < ApplicationController
   before_action :set_relations, only: [:index, :new, :create]
 
   require 'time'
-  require 'fileutils'
 
   # GET /shifts
   # GET /shifts.json
@@ -18,62 +17,6 @@ class ShiftsController < ApplicationController
     @start_date, @end_date = Shift.get_month_range(@search_params[:month])
     @store = Store.find(@search_params[:store_id])
     @staff = Staff.find(@search_params[:staff_id])
-  end
-
-  def new
-    @shift = Shift.new
-  end
-
-  def save
-    uploaded_file = csv_params[:shift_csv]
-    # validate only csv
-    if uploaded_file.content_type != 'text/csv'
-      flash[:alert] = "csvファイルをアップロードしてください。"
-      return redirect_to action: :new
-    end
-
-    file_name = uploaded_file.original_filename
-    save_path = Shift.save_csv_path(file_name)
-    FileUtils.mv(uploaded_file.path, save_path)
-    return redirect_to confirm_shifts_path(file_name: file_name)
-  end
-
-  def confirm
-    @file_name = confirm_params[:file_name]
-    # csvからshiftの配列を作成
-    csv_shifts = Shift.make_from_csv(@file_name)
-    return redirect_to action: :new if csv_shifts.empty?
-
-    # csvのshift全体から最小日付、最大日付を取得
-    @start_date = csv_shifts.min { |s| s.date }&.date
-    @end_date = csv_shifts.max { |s| s.date }&.date
-
-    # paramから検索
-    store_id = confirm_params[:store_id].present? ? confirm_params[:store_id].to_i : csv_shifts.first&.store_id
-    staff_id = confirm_params[:staff_id].present? ? confirm_params[:staff_id].to_i : csv_shifts.first&.staff_id
-    searched = csv_shifts.select { |s| s.store_id == store_id && s.staff_id == staff_id }
-    # 検索したものを日付でグループ化
-    @shifts = searched.group_by(&:date)
-    @store = Store.find(store_id)
-    @staff = Staff.find(staff_id)
-  end
-
-  # POST /shifts
-  # POST /shifts.json
-  def create
-    @file_name = confirm_params[:file_name]
-    return redirect_to action: :new if @file_name.blank?
-
-    begin
-      Shift.transaction do
-        Shift.import(@file_name)
-      end
-      redirect_to action: :index
-    rescue Encoding::UndefinedConversionError => e
-    rescue Encoding::InvalidByteSequenceError => e
-      flash[:alert] = "文字コードがShift-JISのファイルをアップロードしてください。"
-      return redirect_to action: :new
-    end
   end
 
   def updates
@@ -134,14 +77,6 @@ class ShiftsController < ApplicationController
       store_id: params[:store_id] || current_staff.stores.first.id,
       month: search_month
     }
-  end
-
-  def csv_params
-    params.require(:shift).permit(:shift_csv)
-  end
-
-  def confirm_params
-    params.permit(:file_name, :store_id, :staff_id)
   end
 
   def updates_params
