+++ src/app/views/csv_shifts/confirm.html.erb
@@ -1,6 +1,6 @@
 <h1>シフト確認画面</h1>
 
-<%= form_with url: shifts_path(file_name: @file_name) do |form| %>
+<%= form_with url: csv_shifts_path(file_name: @file_name) do |form| %>
   <div class="actions">
     <%= form.submit "登録" %>
   </div>
