+++ src/app/views/csv_shifts/new.html.erb
@@ -1,6 +1,6 @@
 <h1>シフト登録</h1>
 
-<%= form_for @shift, url: save_shifts_path do |form| %>
+<%= form_for @shift, url: save_csv_shifts_path do |form| %>
   <div class="field">
     <%= form.file_field :shift_csv, required: true %>
   </div>
